+++ apptentive/src/main/java/com/apptentive/android/sdk/ApptentiveInternal.java
@@ -42,7 +42,6 @@ import com.apptentive.android.sdk.module.engagement.EngagementModule;
 import com.apptentive.android.sdk.module.engagement.interaction.InteractionManager;
 import com.apptentive.android.sdk.module.engagement.interaction.model.MessageCenterInteraction;
 import com.apptentive.android.sdk.module.messagecenter.MessageManager;
-import com.apptentive.android.sdk.module.metric.MetricModule;
 import com.apptentive.android.sdk.module.rating.IRatingProvider;
 import com.apptentive.android.sdk.module.rating.impl.GooglePlayRatingProvider;
 import com.apptentive.android.sdk.module.survey.OnSurveyFinishedListener;
@@ -153,7 +152,7 @@ public class ApptentiveInternal implements ApptentiveInstance, ApptentiveNotific
 		appRelease = null;
 	}
 
-	private ApptentiveInternal(Application application, ApptentiveConfiguration configuration) {
+	private ApptentiveInternal(Application application, ApptentiveConfiguration configuration, @NonNull String androidID) {
 		if (configuration == null) {
 			throw new IllegalArgumentException("Configuration is null");
 		}
@@ -174,7 +173,8 @@ public class ApptentiveInternal implements ApptentiveInstance, ApptentiveNotific
 		globalSharedPrefs = application.getSharedPreferences(Constants.PREF_NAME, Context.MODE_PRIVATE);
 		apptentiveHttpClient = new ApptentiveHttpClient(apptentiveKey, apptentiveSignature, getEndpointBase(globalSharedPrefs));
 
-		conversationManager = new ConversationManager(appContext, Util.getInternalDir(appContext, CONVERSATIONS_DIR, true), encryption);
+		DeviceManager deviceManager = new DeviceManager(androidID);
+		conversationManager = new ConversationManager(appContext, Util.getInternalDir(appContext, CONVERSATIONS_DIR, true), encryption, deviceManager);
 
 		appRelease = AppReleaseManager.generateCurrentAppRelease(application, this);
 		taskManager = new ApptentiveTaskManager(appContext, apptentiveHttpClient, encryption);
@@ -232,17 +232,20 @@ public class ApptentiveInternal implements ApptentiveInstance, ApptentiveNotific
 
 		synchronized (ApptentiveInternal.class) {
 			if (sApptentiveInternal == null) {
-					ApptentiveLog.i("Registering Apptentive Android SDK %s", Constants.getApptentiveSdkVersion());
-					ApptentiveLog.v("ApptentiveKey=%s ApptentiveSignature=%s", apptentiveKey, apptentiveSignature);
-					sApptentiveInternal = new ApptentiveInternal(application, configuration);
-					dispatchOnConversationQueue(new DispatchTask() {
-						@Override
-						protected void execute() {
-							sApptentiveInternal.start();
-						}
-					});
+				ApptentiveLog.i("Registering Apptentive Android SDK %s", Constants.getApptentiveSdkVersion());
+				ApptentiveLog.v("ApptentiveKey=%s ApptentiveSignature=%s", apptentiveKey, apptentiveSignature);
+				// resolve Android ID
+				boolean shouldGenerateRandomAndroidID = Build.VERSION.SDK_INT < Build.VERSION_CODES.O && !configuration.shouldCollectAndroidIdOnPreOreoTargets();
+				String androidID = resolveAndroidID(application.getApplicationContext(), shouldGenerateRandomAndroidID);
+				sApptentiveInternal = new ApptentiveInternal(application, configuration, androidID);
+				dispatchOnConversationQueue(new DispatchTask() {
+					@Override
+					protected void execute() {
+						sApptentiveInternal.start();
+					}
+				});
 
-					ApptentiveActivityLifecycleCallbacks.register(application);
+				ApptentiveActivityLifecycleCallbacks.register(application);
 			} else {
 				ApptentiveLog.w("Apptentive instance is already initialized");
 			}
@@ -1188,6 +1191,38 @@ public class ApptentiveInternal implements ApptentiveInstance, ApptentiveNotific
 
 	private static void logException(Exception e) {
 		ErrorMetrics.logException(e); // TODO: add more context info
+	}
+
+	//endregion
+
+	//region Android ID
+
+	private static final String PREFS_NAME_ANDROID_ID = "com.apptentive.sdk.androidID";
+	private static final String PREFS_KEY_NAME_ANDROID_ID = "androidID";
+
+	private static String resolveAndroidID(Context context, boolean shouldGenerateRandomAndroidID) {
+		if (shouldGenerateRandomAndroidID) {
+			String existingAndroidID = loadAndroidID(context);
+			if (existingAndroidID != null) {
+				return existingAndroidID;
+			}
+
+			String androidID = StringUtils.randomAndroidID();
+			saveAndroidID(context, androidID);
+			return androidID;
+		}
+
+		return Util.getAndroidID(context);
+	}
+
+	private static String loadAndroidID(Context context) {
+		SharedPreferences sharedPreferences = context.getSharedPreferences(PREFS_NAME_ANDROID_ID, Context.MODE_PRIVATE);
+		return sharedPreferences.getString(PREFS_KEY_NAME_ANDROID_ID, null);
+	}
+
+	private static void saveAndroidID(Context context, String androidID) {
+		SharedPreferences sharedPreferences = context.getSharedPreferences(PREFS_NAME_ANDROID_ID, Context.MODE_PRIVATE);
+		sharedPreferences.edit().putString(PREFS_KEY_NAME_ANDROID_ID, androidID).apply();
 	}
 
 	//endregion
