+++ apptentive/src/main/java/com/apptentive/android/sdk/storage/DeviceManager.java
@@ -12,21 +12,26 @@ import android.telephony.TelephonyManager;
 
 import com.apptentive.android.sdk.ApptentiveLog;
 import com.apptentive.android.sdk.model.Configuration;
-import com.apptentive.android.sdk.model.DevicePayload;
 import com.apptentive.android.sdk.util.AdvertiserManager;
 import com.apptentive.android.sdk.util.Constants;
-import com.apptentive.android.sdk.util.Util;
+import com.apptentive.android.sdk.util.StringUtils;
 
 import java.util.Locale;
 import java.util.TimeZone;
 
 import static com.apptentive.android.sdk.debug.ErrorMetrics.logException;
 
-/**
- * A helper class with static methods for and diffing information about the current device.
- */
 public class DeviceManager {
-	public static Device generateNewDevice(Context context) {
+	private final String androidID;
+
+	public DeviceManager(String androidID) {
+		if (StringUtils.isNullOrEmpty(androidID)) {
+			throw new IllegalArgumentException("Android ID is null or empty");
+		}
+		this.androidID = androidID;
+	}
+
+	public Device generateNewDevice(Context context) {
 		Device device = new Device();
 
 		// First, get all the information we can load from static resources.
@@ -41,7 +46,7 @@ public class DeviceManager {
 		device.setBrand(Build.BRAND);
 		device.setCpu(Build.CPU_ABI);
 		device.setDevice(Build.DEVICE);
-		device.setUuid(Util.getAndroidId(context));
+		device.setUuid(androidID);
 		device.setBuildType(Build.TYPE);
 		device.setBuildId(Build.ID);
 
@@ -82,150 +87,5 @@ public class DeviceManager {
 		device.setUtcOffset(String.valueOf((TimeZone.getDefault().getRawOffset() / 1000)));
 		return device;
 	}
-
-	public static DevicePayload getDiffPayload(com.apptentive.android.sdk.storage.Device oldDevice, com.apptentive.android.sdk.storage.Device newDevice) {
-		if (newDevice == null) {
-			return null;
-		}
-
-		DevicePayload ret = new DevicePayload();
-		boolean changed = false;
-
-		if (oldDevice == null || !equal(oldDevice.getUuid(), newDevice.getUuid())) {
-			ret.setUuid(newDevice.getUuid());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getOsName(), newDevice.getOsName())) {
-			ret.setOsName(newDevice.getOsName());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getOsVersion(), newDevice.getOsVersion())) {
-			ret.setOsVersion(newDevice.getOsVersion());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getOsBuild(), newDevice.getOsBuild())) {
-			ret.setOsBuild(newDevice.getOsBuild());
-			changed = true;
-		}
-
-		if (oldDevice == null || oldDevice.getOsApiLevel() != newDevice.getOsApiLevel()) {
-			ret.setOsApiLevel(String.valueOf(newDevice.getOsApiLevel()));
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getManufacturer(), newDevice.getManufacturer())) {
-			ret.setManufacturer(newDevice.getManufacturer());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getModel(), newDevice.getModel())) {
-			ret.setModel(newDevice.getModel());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getBoard(), newDevice.getBoard())) {
-			ret.setBoard(newDevice.getBoard());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getProduct(), newDevice.getProduct())) {
-			ret.setProduct(newDevice.getProduct());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getBrand(), newDevice.getBrand())) {
-			ret.setBrand(newDevice.getBrand());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getCpu(), newDevice.getCpu())) {
-			ret.setCpu(newDevice.getCpu());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getDevice(), newDevice.getDevice())) {
-			ret.setDevice(newDevice.getDevice());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getCarrier(), newDevice.getCarrier())) {
-			ret.setCarrier(newDevice.getCarrier());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getCurrentCarrier(), newDevice.getCurrentCarrier())) {
-			ret.setCurrentCarrier(newDevice.getCurrentCarrier());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getNetworkType(), newDevice.getNetworkType())) {
-			ret.setNetworkType(newDevice.getNetworkType());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getBuildType(), newDevice.getBuildType())) {
-			ret.setBuildType(newDevice.getBuildType());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getBuildId(), newDevice.getBuildId())) {
-			ret.setBuildId(newDevice.getBuildId());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getBootloaderVersion(), newDevice.getBootloaderVersion())) {
-			ret.setBootloaderVersion(newDevice.getBootloaderVersion());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getRadioVersion(), newDevice.getRadioVersion())) {
-			ret.setRadioVersion(newDevice.getRadioVersion());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getCustomData(), newDevice.getCustomData())) {
-			CustomData customData = newDevice.getCustomData();
-			ret.setCustomData(customData != null ? customData.toJson() : null);
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getLocaleCountryCode(), newDevice.getLocaleCountryCode())) {
-			ret.setLocaleCountryCode(newDevice.getLocaleCountryCode());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getLocaleLanguageCode(), newDevice.getLocaleLanguageCode())) {
-			ret.setLocaleLanguageCode(newDevice.getLocaleLanguageCode());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getLocaleRaw(), newDevice.getLocaleRaw())) {
-			ret.setLocaleRaw(newDevice.getLocaleRaw());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getUtcOffset(), newDevice.getUtcOffset())) {
-			ret.setUtcOffset(newDevice.getUtcOffset());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getAdvertiserId(), newDevice.getAdvertiserId())) {
-			ret.setAdvertiserId(newDevice.getAdvertiserId());
-			changed = true;
-		}
-
-		if (oldDevice == null || !equal(oldDevice.getIntegrationConfig(), newDevice.getIntegrationConfig())) {
-			IntegrationConfig integrationConfig = newDevice.getIntegrationConfig();
-			ret.setIntegrationConfig(integrationConfig != null ? integrationConfig.toJson() : null);
-			changed = true;
-		}
-		return changed ? ret : null;
-	}
-
-	private static boolean equal(Object a, Object b) {
-		return a == null && b == null || a != null && b != null && a.equals(b);
-	}
 }
+
