+++ apptentive/src/main/java/com/apptentive/android/sdk/util/Constants.java
@@ -9,7 +9,7 @@ package com.apptentive.android.sdk.util;
 public class Constants {
 
 	public static final int  API_VERSION = 9;
-	private static final String APPTENTIVE_SDK_VERSION = "5.4.2";
+	private static final String APPTENTIVE_SDK_VERSION = "5.4.3";
 
 	public static final int DEFAULT_CONNECT_TIMEOUT_MILLIS = 45000;
 	public static final int DEFAULT_READ_TIMEOUT_MILLIS = 45000;
